Post-review changes; moved components to the admin directory
