+++ src/components/admin/Header/CrubmtrailSeparatorIcon.tsx
@@ -1,0 +1,11 @@
+export const CrumbtrailSeparatorIcon: React.FC = () => (
+  <svg
+    className="ml-2 flex-shrink-0 h-5 w-5 text-gray-400"
+    xmlns="http://www.w3.org/2000/svg"
+    fill="currentColor"
+    viewBox="0 0 20 20"
+    aria-hidden="true"
+  >
+    <path d="M5.555 17.776l8-16 .894.448-8 16-.894-.448z" />
+  </svg>
+);
