--- src/components/admin/Header/Header.tailwind.ts
@@ -1,6 +1,0 @@
-export const breadcrumb =
-  "cursor-pointer text-sm font-medium text-gray-500 hover:text-gray-700";
-export const title =
-  "mt-2 text-3xl font-medium leading-7 text-gray-900 sm:text-4xl sm:truncate";
-export const subtitle = "mt-2 flex items-center text-sm text-gray-500";
-export const row = "flex row";
