+++ src/components/admin/ImageInput/ImageInput.tailwind.ts
@@ -1,0 +1,4 @@
+export const wrapper = "justify-center items-center flex bg-cover";
+export const wideWrapper = "w-full text-white rounded-md";
+export const roundWrapper =
+  "w-48 h-48 text-white rounded-full flex-shrink-0 mx-auto";
