--- src/components/atoms/AdminHeader/index.tsx
@@ -1,1 +1,0 @@
-export { AdminHeader } from "./AdminHeader";
