+++ src/pages/WorldEditor/WorldEditorGeneralPanel.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { ThreeColumnLayout } from "components/admin/ThreeColumnLayout";
+import { WorldGeneralForm } from "components/admin/WorldGeneralForm";
+import { WorldShowcase } from "components/admin/WorldShowcase";
 
 import { useWorldBySlug } from "hooks/worlds/useWorldBySlug";
 
 import { AdminPanel } from "components/organisms/AdminVenueView/components/AdminPanel";
 import { AdminShowcase } from "components/organisms/AdminVenueView/components/AdminShowcase";
 import { AdminSidebar } from "components/organisms/AdminVenueView/components/AdminSidebar";
-import { WorldGeneralForm } from "components/organisms/WorldGeneralForm";
-import { WorldShowcase } from "components/organisms/WorldShowcase";
 
 import { Loading } from "components/molecules/Loading";
