+++ src/settings/adminDesignSettings.ts
@@ -1,0 +1,16 @@
+import { STRING_COPYRIGHT } from "./interpolationSettings";
+
+export const CURRENT_YEAR = new Date().getFullYear();
+
+export const COPYRIGHT_TEXT = `${STRING_COPYRIGHT} Copyright SparkleVerse Inc & Contributors 2020 to ${CURRENT_YEAR}.`;
+
+export const SPACE_CARD_HEADER_BACKGROUND =
+  "https://images.unsplash.com/photo-1517369611053-aecb0ca6971e?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=987&q=80";
+
+export const DEFAULT_IMAGE_INPUT_BACKGROUND =
+  "https://media.istockphoto.com/photos/blue-landscape-of-glowing-particles-picture-id1287065311?b=1&k=20&m=1287065311&s=170667a&w=0&h=04F0a32sm82f12MMeluipyUFkW562H6sCEEmAmGzPwo=";
+
+export const DEFAULT_IMAGE_INPUT_ROUND_BACKGOUND =
+  "https://images.unsplash.com/photo-1540957904852-c2184978df5e?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=2070&q=80";
+
+export const DEFAULT_TEXTAREA_ROWS = 4;
