+++ src/settings/index.ts
@@ -5,7 +5,6 @@ import { VenueTemplate } from "types/VenueTemplate";
 import { generateTables } from "utils/table";
 import { FIVE_MINUTES_MS } from "utils/time";
 
-import { STRING_COPYRIGHT } from "./interpolationSettings";
 import {
   ROOM_TAXON,
   ROOMS_TAXON,
@@ -15,6 +14,7 @@ import {
 } from "./taxonomy";
 
 // NOTE: please keep these exports sorted alphabetically for faster visual scan
+export * from "./adminDesignSettings";
 export * from "./apiSettings";
 export * from "./dateSettings";
 export * from "./disableSettings";
@@ -109,8 +109,6 @@ export const MAXIMUM_AUDITORIUM_COLUMNS_COUNT = 5;
 export const MINIMUM_AUDITORIUM_ROWS_COUNT = 5;
 export const MAXIMUM_AUDITORIUM_ROWS_COUNT = 5;
 // playa is 4000x4000 pixels, Burning Seed paddock is 2000x2000
-
-export const DEFAULT_TEXTAREA_ROWS = 4;
 
 export const ACCEPTED_IMAGE_TYPES =
   "image/png,image/x-png,image/gif,image/jpg,image/jpeg,image/tiff,image/bmp,image/gif,image/webp";
@@ -615,15 +613,3 @@ export const DEFAULT_SHOW_MORE_SETTINGS = {
   truncatedEndingComponent: "... ",
 };
 export const DEFAULT_MISSING_PLACEHOLDER = "Placeholder";
-
-export const CURRENT_YEAR = new Date().getFullYear();
-export const COPYRIGHT_TEXT = `${STRING_COPYRIGHT} Copyright SparkleVerse Inc & Contributors 2020 to ${CURRENT_YEAR}.`;
-
-export const SPACE_CARD_HEADER_BACKGROUND =
-  "https://images.unsplash.com/photo-1517369611053-aecb0ca6971e?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=987&q=80";
-
-export const DEFAULT_IMAGE_INPUT_BACKGROUND =
-  "https://media.istockphoto.com/photos/blue-landscape-of-glowing-particles-picture-id1287065311?b=1&k=20&m=1287065311&s=170667a&w=0&h=04F0a32sm82f12MMeluipyUFkW562H6sCEEmAmGzPwo=";
-
-export const DEFAULT_IMAGE_INPUT_ROUND_BACKGOUND =
-  "https://images.unsplash.com/photo-1540957904852-c2184978df5e?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=2070&q=80";
